requirements.txt
Django==4.2.7
Pillow>=10.0.0
django-ckeditor>=6.0.0
django-cleanup>=8.0.0
python-decouple>=3.8
django-crispy-forms>=2.0
crispy-tailwind>=0.5.0
django-taggit>=4.0.0
django-simple-history>=3.4.0
django-extensions>=3.2.0
